Migrated initial batch of functions to ES6 modules
Converted about 30% of the function modules into native ES6 modules. Modules that had  calls are now exported as the default in an object format. The actual registration calls themselves are now handled in . For modules that share functionality such as constructors or functions, these too are expected as named exports and can be imported via . Also verified that  imports are indeed singletons, so each import isn't reinstantiating the  module. Found a couple circular dependencies that will need to be investigated, but for now I'm marching onwards.