Completed first major ES6 refactor
Each function module is now an ES6 module that emits a "registration" object that 'src/higher.js' uses to do it's registration magic. The default export however is now the "implementation" of the function itself. This makes importing the function in another module pure -- no context is needed and all you are getting when you import is the function itself.

The 'register' method has been refactored to combine the "alias" function. To register an alias now, just include "aliases" property on the registration object, and the registered method will also link an alias using the provided names.